Work in progress - committing to back up my work

+++ src/libc/regex/TRE/lib/tre-mem.c
@@ -106,12 +106,12 @@ tre_mem_alloc_impl(tre_mem_t mem, int provided, void *provided_block,
 	}
       else
 	{
-	  int block_size;
+	  size_t block_size;
 	  if (size * 8 > TRE_MEM_BLOCK_SIZE)
 	    block_size = size * 8;
 	  else
 	    block_size = TRE_MEM_BLOCK_SIZE;
-	  DPRINT(("tre_mem_alloc: allocating new %d byte block\n",
+	  DPRINT(("tre_mem_alloc: allocating new %ld byte block\n",
 		  block_size));
 	  l = xmalloc(sizeof(*l));
 	  if (l == NULL)
